added pagination and improved component's folder structure
